refactor: remove localhost client state, condense localhost impl into single file
