--- modules/light-clients/09-localhost/client_state.go
@@ -1,151 +1,0 @@
-package localhost
-
-import (
-	"bytes"
-
-	errorsmod "cosmossdk.io/errors"
-	storetypes "cosmossdk.io/store/types"
-
-	"github.com/cosmos/cosmos-sdk/codec"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	clienttypes "github.com/cosmos/ibc-go/v8/modules/core/02-client/types"
-	commitmenttypes "github.com/cosmos/ibc-go/v8/modules/core/23-commitment/types"
-	host "github.com/cosmos/ibc-go/v8/modules/core/24-host"
-	ibcerrors "github.com/cosmos/ibc-go/v8/modules/core/errors"
-	"github.com/cosmos/ibc-go/v8/modules/core/exported"
-)
-
-var _ exported.ClientState = (*ClientState)(nil)
-
-// NewClientState creates a new 09-localhost ClientState instance.
-func NewClientState(height clienttypes.Height) *ClientState {
-	return &ClientState{
-		LatestHeight: height,
-	}
-}
-
-// ClientType returns the 09-localhost client type.
-func (ClientState) ClientType() string {
-	return exported.Localhost
-}
-
-// Validate performs a basic validation of the client state fields.
-func (cs ClientState) Validate() error {
-	if cs.LatestHeight.RevisionHeight == 0 {
-		return errorsmod.Wrapf(ibcerrors.ErrInvalidHeight, "local revision height cannot be zero")
-	}
-
-	return nil
-}
-
-// Initialize ensures that initial consensus state for localhost is nil.
-func (ClientState) Initialize(ctx sdk.Context, cdc codec.BinaryCodec, clientStore storetypes.KVStore, consState exported.ConsensusState) error {
-	if consState != nil {
-		return errorsmod.Wrap(clienttypes.ErrInvalidConsensus, "initial consensus state for localhost must be nil.")
-	}
-
-	return nil
-}
-
-// GetTimestampAtHeight returns the current block time retrieved from the application context. The localhost client does not store consensus states and thus
-// cannot provide a timestamp for the provided height.
-func (ClientState) GetTimestampAtHeight(ctx sdk.Context, _ storetypes.KVStore, _ codec.BinaryCodec, _ exported.Height) (uint64, error) {
-	return uint64(ctx.BlockTime().UnixNano()), nil
-}
-
-// VerifyMembership is a generic proof verification method which verifies the existence of a given key and value within the IBC store.
-// The caller is expected to construct the full CommitmentPath from a CommitmentPrefix and a standardized path (as defined in ICS 24).
-// The caller must provide the full IBC store.
-func (ClientState) VerifyMembership(
-	ctx sdk.Context,
-	store storetypes.KVStore,
-	_ codec.BinaryCodec,
-	_ exported.Height,
-	_ uint64,
-	_ uint64,
-	proof []byte,
-	path exported.Path,
-	value []byte,
-) error {
-	// ensure the proof provided is the expected sentinel localhost client proof
-	if !bytes.Equal(proof, SentinelProof) {
-		return errorsmod.Wrapf(commitmenttypes.ErrInvalidProof, "expected %s, got %s", string(SentinelProof), string(proof))
-	}
-
-	merklePath, ok := path.(commitmenttypes.MerklePath)
-	if !ok {
-		return errorsmod.Wrapf(ibcerrors.ErrInvalidType, "expected %T, got %T", commitmenttypes.MerklePath{}, path)
-	}
-
-	if len(merklePath.GetKeyPath()) != 2 {
-		return errorsmod.Wrapf(host.ErrInvalidPath, "path must be of length 2: %s", merklePath.GetKeyPath())
-	}
-
-	// The commitment prefix (eg: "ibc") is omitted when operating on the core IBC store
-	bz := store.Get([]byte(merklePath.KeyPath[1]))
-	if bz == nil {
-		return errorsmod.Wrapf(clienttypes.ErrFailedMembershipVerification, "value not found for path %s", path)
-	}
-
-	if !bytes.Equal(bz, value) {
-		return errorsmod.Wrapf(clienttypes.ErrFailedMembershipVerification, "value provided does not equal value stored at path: %s", path)
-	}
-
-	return nil
-}
-
-// VerifyNonMembership is a generic proof verification method which verifies the absence of a given CommitmentPath within the IBC store.
-// The caller is expected to construct the full CommitmentPath from a CommitmentPrefix and a standardized path (as defined in ICS 24).
-// The caller must provide the full IBC store.
-func (ClientState) VerifyNonMembership(
-	ctx sdk.Context,
-	store storetypes.KVStore,
-	_ codec.BinaryCodec,
-	_ exported.Height,
-	_ uint64,
-	_ uint64,
-	proof []byte,
-	path exported.Path,
-) error {
-	// ensure the proof provided is the expected sentinel localhost client proof
-	if !bytes.Equal(proof, SentinelProof) {
-		return errorsmod.Wrapf(commitmenttypes.ErrInvalidProof, "expected %s, got %s", string(SentinelProof), string(proof))
-	}
-
-	merklePath, ok := path.(commitmenttypes.MerklePath)
-	if !ok {
-		return errorsmod.Wrapf(ibcerrors.ErrInvalidType, "expected %T, got %T", commitmenttypes.MerklePath{}, path)
-	}
-
-	if len(merklePath.GetKeyPath()) != 2 {
-		return errorsmod.Wrapf(host.ErrInvalidPath, "path must be of length 2: %s", merklePath.GetKeyPath())
-	}
-
-	// The commitment prefix (eg: "ibc") is omitted when operating on the core IBC store
-	if store.Has([]byte(merklePath.KeyPath[1])) {
-		return errorsmod.Wrapf(clienttypes.ErrFailedNonMembershipVerification, "value found for path %s", path)
-	}
-
-	return nil
-}
-
-// VerifyClientMessage is unsupported by the 09-localhost client type and returns an error.
-func (ClientState) VerifyClientMessage(_ sdk.Context, _ codec.BinaryCodec, _ storetypes.KVStore, _ exported.ClientMessage) error {
-	return errorsmod.Wrap(clienttypes.ErrUpdateClientFailed, "client message verification is unsupported by the localhost client")
-}
-
-// CheckForMisbehaviour is unsupported by the 09-localhost client type and performs a no-op, returning false.
-func (ClientState) CheckForMisbehaviour(_ sdk.Context, _ codec.BinaryCodec, _ storetypes.KVStore, _ exported.ClientMessage) bool {
-	return false
-}
-
-// UpdateStateOnMisbehaviour is unsupported by the 09-localhost client type and performs a no-op.
-func (ClientState) UpdateStateOnMisbehaviour(_ sdk.Context, _ codec.BinaryCodec, _ storetypes.KVStore, _ exported.ClientMessage) {
-}
-
-// UpdateState updates and stores as necessary any associated information for an IBC client, such as the ClientState and corresponding ConsensusState.
-// Upon successful update, a list of consensus heights is returned. It assumes the ClientMessage has already been verified.
-func (ClientState) UpdateState(ctx sdk.Context, _ codec.BinaryCodec, _ storetypes.KVStore, _ exported.ClientMessage) []exported.Height {
-	return []exported.Height{clienttypes.GetSelfHeight(ctx)}
-}
