--- modules/light-clients/09-localhost/codec.go
@@ -1,16 +1,0 @@
-package localhost
-
-import (
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-
-	"github.com/cosmos/ibc-go/v8/modules/core/exported"
-)
-
-// RegisterInterfaces registers the tendermint concrete client-related
-// implementations and interfaces.
-func RegisterInterfaces(registry codectypes.InterfaceRegistry) {
-	registry.RegisterImplementations(
-		(*exported.ClientState)(nil),
-		&ClientState{},
-	)
-}
