--- modules/light-clients/09-localhost/codec_test.go
@@ -1,49 +1,0 @@
-package localhost_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
-
-	ibc "github.com/cosmos/ibc-go/v8/modules/core"
-	localhost "github.com/cosmos/ibc-go/v8/modules/light-clients/09-localhost"
-)
-
-func TestCodecTypeRegistration(t *testing.T) {
-	testCases := []struct {
-		name    string
-		typeURL string
-		expPass bool
-	}{
-		{
-			"success: ClientState",
-			sdk.MsgTypeURL(&localhost.ClientState{}),
-			true,
-		},
-		{
-			"type not registered on codec",
-			"ibc.invalid.MsgTypeURL",
-			false,
-		},
-	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		t.Run(tc.name, func(t *testing.T) {
-			encodingCfg := moduletestutil.MakeTestEncodingConfig(ibc.AppModuleBasic{})
-			msg, err := encodingCfg.Codec.InterfaceRegistry().Resolve(tc.typeURL)
-
-			if tc.expPass {
-				require.NotNil(t, msg)
-				require.NoError(t, err)
-			} else {
-				require.Nil(t, msg)
-				require.Error(t, err)
-			}
-		})
-	}
-}
