--- proto/ibc/lightclients/localhost/v2/localhost.proto
@@ -1,16 +1,0 @@
-syntax = "proto3";
-
-package ibc.lightclients.localhost.v2;
-
-option go_package = "github.com/cosmos/ibc-go/v8/modules/light-clients/09-localhost;localhost";
-
-import "ibc/core/client/v1/client.proto";
-import "gogoproto/gogo.proto";
-
-// ClientState defines the 09-localhost client state
-message ClientState {
-  option (gogoproto.goproto_getters) = false;
-
-  // the latest block height
-  ibc.core.client.v1.Height latest_height = 1 [(gogoproto.nullable) = false];
-}
